Weight sampler class for sampling from Dirichlet dist. Refactor Bayesian weight learning. Remove scaling option in CRGS. Remove mean and variance of weight sample in CRGS and Hyperband. Remove hypersphere option. Remove log scaling options.
Weight sampler class for sampling from Dirichlet dist.

Remove scaling option in CRGS, this is not useful as is shown in the MLJ2020 paper.

Remove mean and variance of weight sample in CRGS and Hyperband. Scaling and biasing the sample in CRGS and Hyperband was initially added with the intent of sampling from a non-standard gaussian, as implied by original comments. However the sampling was not from a standard gaussian it was from a uniform distribution over [0,1]. This had the effect of sampling from cube in the weight space.

Remove hypersphere option.
This can simply be approximated with the Dirichlet sampling method.

 Remove log scaling options.