Allow atom managers to return null in getAtom calls.
This is a part of the changes made in the larger onlinePSL project.

Allow atom managers to return null in getAtom calls.

This can occur when onlineAtom managers decide the atom is invalid but should not throw an access exception. This happens when adding atoms since adding atoms should not create new random variable atoms.